finished upto part 4.23

+++ part4/controllers/blogs.js
@@ -1,14 +1,10 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
+const jwt = require('jsonwebtoken');
 
-// For testing will remove later
 
-blogsRouter.get('/', (request, response) => {
-    response.send('Blog List backend.');
-});
-
-blogsRouter.get('/api/blogs', async (request, response) => {
+blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
         .find({})
         .populate('user', { username: 1, name: 1 });
@@ -16,7 +12,7 @@ blogsRouter.get('/api/blogs', async (request, response) => {
     response.json(blogs);
 });
 
-blogsRouter.get('/api/blogs/:id', async (request, response) => {
+blogsRouter.get('/:id', async (request, response) => {
     const blog = await Blog.findById(request.params.id);
 
     if (blog) {
@@ -26,11 +22,15 @@ blogsRouter.get('/api/blogs/:id', async (request, response) => {
     }
 });
 
-blogsRouter.post('/api/blogs', async (request, response) => {
+blogsRouter.post('/', async (request, response) => {
     const body = request.body;
+    const user = request.user;
+    const token = request.token;
 
-    const users = await User.find({});
-    const user = users[0];
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    if (!(token && decodedToken.id)) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
 
     const blog = new Blog(
         {
@@ -50,13 +50,31 @@ blogsRouter.post('/api/blogs', async (request, response) => {
     response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete('/api/blogs/:id', async (request, response) => {
+blogsRouter.delete('/:id', async (request, response) => {
+    const user = request.user;
+    const blog = await Blog.findById(request.params.id);
+
+    if (!blog) {
+        return response.status(404).json({
+            error: 'blog does not exist'
+        });
+    } else if (!user) {
+        return response.status(401).json({
+            error: 'unathorized request'
+        });
+    }
+    
+    // Check to make sure it is user who posted blog trying to delete
+    if (blog.user._id.toString() !== user._id.toString()) {
+        return response.status(401).json({ error: 'blogs can only be deleted by publisher' });
+    }
+
     await Blog.findByIdAndRemove(request.params.id);
 
     response.status(204).end();
 });
 
-blogsRouter.put('/api/blogs/:id', async (request, response) => {
+blogsRouter.put('/:id', async (request, response) => {
     const likes = request.body.likes;
     
     // update the likes
